requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.1
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.30
langsmith==0.2.11
oauthlib==3.2.2
openai==1.59.9
orjson==3.10.15
packaging==24.2
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==9.0.0
tqdm==4.67.1
tweepy==4.15.0
typing_extensions==4.12.2
urllib3==2.3.0
